+++ script/build-resources.ts
@@ -2,8 +2,8 @@
  * build-resources.ts: This script obtains the resources needed to build the
  * project, such as downloading the latest version of ANTLR.
  */
-import * as fs from 'node:fs';
-import * as path from 'node:path';
+import fs from 'node:fs';
+import path from 'node:path';
 import getHtmlParsed from './helper/getHtmlParsed';
 import downloadIfNotExist from './helper/downloadIfNotExist';
 
@@ -29,10 +29,13 @@ const resourceDir = 'res';
         ).href;
     const antlrPath = path.resolve(__dirname, '..', resourceDir, 'antlr-complete.jar');
     const antlrVersionPath = path.resolve(__dirname, '..', resourceDir, 'antlr-version.txt');
+    const antlrLicenseUrl = 'https://raw.githubusercontent.com/antlr/antlr4/refs/heads/dev/LICENSE.txt';
+    const antlrLicensePath = path.resolve(__dirname, '..', resourceDir, 'antlr-LICENSE.txt');
     let antlrInstalledVersion: string;
     try {
         fs.accessSync(antlrPath, fs.constants.R_OK);
         fs.accessSync(antlrVersionPath, fs.constants.R_OK);
+        fs.accessSync(antlrLicensePath, fs.constants.R_OK);
         antlrInstalledVersion = fs.readFileSync(antlrVersionPath, 'utf-8').trim();
     } catch {
         antlrInstalledVersion = '0.0.0';
@@ -44,6 +47,7 @@ const resourceDir = 'res';
         } catch {}
         console.log(`Downloading ANTLR version ${antlrVersion} ...`);
         await downloadIfNotExist(antlrUrl, antlrPath);
+        await downloadIfNotExist(antlrLicenseUrl, antlrLicensePath);
         fs.writeFileSync(antlrVersionPath, antlrVersion);
         console.log(`Downloading ANTLR version ${antlrVersion} done.`);
     } else {