+++ src/CharString.spec.ts
@@ -1,7 +1,12 @@
+import path from 'node:path';
 import { CharString } from './CharString';
 
-describe('CharString', () => {
-    it('CharString should be defined', () => {
+const __filenameMatch = __filename.match(new RegExp(`.*\\${path.sep}([^\\${path.sep}]+)\\.spec\\.([cm]?[jt]s)\$`))!;
+const unitName = __filenameMatch[1];
+const testExtension = __filenameMatch[2];
+
+describe(`${unitName} unit test (.${testExtension} test file).`, () => {
+    it(`${unitName} should be defined.`, () => {
         expect(CharString).toBeDefined();
     });
 });