+++ src/CoreFunctions.spec.ts
@@ -1,7 +1,12 @@
+import path from 'node:path';
 import { CoreFunctions } from './CoreFunctions';
 
-describe('CoreFunctions', () => {
-    it('CoreFunctions and its functions should be defined', () => {
+const __filenameMatch = __filename.match(new RegExp(`.*\\${path.sep}([^\\${path.sep}]+)\\.spec\\.([cm]?[jt]s)\$`))!;
+const unitName = __filenameMatch[1];
+const testExtension = __filenameMatch[2];
+
+describe(`${unitName} unit test (.${testExtension} test file).`, () => {
+    it(`${unitName} and its methods should be defined.`, () => {
         expect(CoreFunctions).toBeDefined();
     });
 });