+++ src/Structure.spec.ts
@@ -1,9 +1,12 @@
+import path from 'node:path';
 import { Structure } from './Structure';
 
-describe('Structure', () => {
-    beforeEach(async () => {});
+const __filenameMatch = __filename.match(new RegExp(`.*\\${path.sep}([^\\${path.sep}]+)\\.spec\\.([cm]?[jt]s)\$`))!;
+const unitName = __filenameMatch[1];
+const testExtension = __filenameMatch[2];
 
-    it('Structure should be defined', () => {
+describe(`${unitName} unit test (.${testExtension} test file).`, () => {
+    it(`${unitName} should be defined.`, () => {
         // expect(Structure).toBeDefined();
     });
 });