+++ src/global.d.ts
@@ -1,0 +1,1 @@
+declare const __WEBPACK_BUNDLE_CONFIGURATION: string;