+++ src/lib-core.ts
@@ -1,0 +1,17 @@
+export { Decimal } from 'decimal.js';
+export { CharString } from './CharString';
+export { ComplexDecimal } from './ComplexDecimal';
+export { MultiArray } from './MultiArray';
+export { Structure } from './Structure';
+export { FunctionHandle } from './FunctionHandle';
+export { LinearAlgebra } from './LinearAlgebra';
+export { MathOperation } from './MathOperation';
+export { CoreFunctions } from './CoreFunctions';
+export { Configuration } from './Configuration';
+export { SymbolTable } from './SymbolTable';
+export * as AST from './AST';
+export * from './Evaluator';
+/**
+ * The following constant is defined in `webpack.config.ts`
+ */
+export const bundleConfiguration = __WEBPACK_BUNDLE_CONFIGURATION;